pyproject.toml
[project]
name = "cafa6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "altair[all]<6.0.0",
    "biopython>=1.86",
    "esm>=3.2.1.post1",
    "flash-attn>=2.8.3",
    "google-genai>=1.52.0",
    "h5py>=3.15.1",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "jax[cuda]>=0.6.2",
    "jupyter>=1.1.1",
    "kaggle>=1.7.4.5",
    "keras>=3.12.0",
    "keras-tuner>=1.4.8",
    "marimo>=0.18.2",
    "nbformat>=5.10.4",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "polars>=1.35.2",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "pyzmq>=27.1.0",
    "scikit-learn>=1.7.2",
    "tensorflow[and-cuda]>=2.20.0",
    "tqdm>=4.67.1",
    "wandb>=0.23.0",
    "xgboost>=3.1.2",
]
